OdevdockerfileVersion1/client_api/requirements.txt
Flask==3.0.2
requests==2.31.0
flask-cors
PyJWT==2.8.0
